Address the fourth Codex review: verify writes, form refresh, stream re-auth, strictness
- vendored write.proto gains the verify operation the Firebase SDKs send for
  transaction reads (documented local patch); WriteOp::Verify checks the precondition
  without changing anything, rules treat it as a get
- token refresh accepts application/x-www-form-urlencoded bodies (the JS SDK's shape)
- streams re-verify the bearer token on every write / refresh and end on a session
  reset (backend epoch); first Write message must be the handshake, acknowledgements
  never regress, inbound errors and closed responses end the tasks
- rules: lossless i128 timestamps and request.time, previews and rules time use the
  commit time the commit will receive, one ruleset snapshot per operation
- REST: listCollectionIds owner-only with paging, percent-decoded path segments, readTime
  selectors and readOnly.readTime carried to the backend (refused explicitly),
  findNearest / unknown server values / malformed geo points rejected
- gRPC decoding limit raised to the Firestore request size; custom-token sign-in records
  lastLoginAt

+++ crates/ftd-core-rules/src/value.rs
@@ -24,8 +24,8 @@ pub enum RulesValue {
     Map(BTreeMap<String, RulesValue>),
     /// Path (segments without the leading slash).
     Path(Vec<String>),
-    /// Timestamp as Unix nanoseconds (full Firestore precision).
-    Timestamp(i64),
+    /// Timestamp as Unix nanoseconds (full Firestore precision, no saturation).
+    Timestamp(i128),
     /// Bytes.
     Bytes(Vec<u8>),
     /// Geo point.
